Fix and update tests Fix warnoings in tests Not a subject for review

+++ src/components/security_manager/src/ssl_context_impl.cc
@@ -208,7 +208,7 @@ bool CryptoManagerImpl::SSLContextImpl::Encrypt(
 
   EnsureBufferSizeEnough(len);
   const int read_size = BIO_read(bioOut_, buffer_, len);
-  DCHECK(len == read_size);
+  DCHECK(len == static_cast<size_t>(read_size));
   if (read_size <= 0) {
     // Reset filter and connection deinitilization instead
     BIO_reset(bioFilter_);
